Fix enum circular imports
- Define relevant enums in cube.ts
- Fix all imports to not be circular
